TF/workload/DLRM-tf/requirements.txt
intel-tensorflow==2.2
tqdm
